All reward buttons are now clickable and shows the right details on the reward detail page. Reward service changed a bit so it gives back the information for the certain event + clicked button Only things to do now are making the buy button work  and saving the reward

+++ ReCrowd/Domain/Reward.swift
@@ -11,9 +11,9 @@ import Foundation
 class Reward: NSObject {
     public var name: String
     public var _description: String
-    public var cost: Double
+    public var cost: Int
     
-    init(named name: String, withDescription description: String, withCost cost: Double) {
+    init(named name: String, withDescription description: String, withCost cost: Int) {
         self.name = name
         self._description = description
         self.cost = cost
